feature: read data from Google wallet and Revolut notifications to create expense items in the background + rewrite to fulfill agent.md requirements

+++ core/data/src/commonMain/kotlin/prayit/simplebudget/core/data/repository/ExpenseRepositoryImpl.kt
@@ -47,6 +47,6 @@ private fun Expense.toEntity() = ExpenseEntity(
     id = id,
     title = title,
     amount = amount,
-    date = date.toEpochDays().toLong(),
+    date = date.toEpochDays(),
     tag = tag,
 )
